+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Download current and previous releases from [https://github.com/ludus-leonis/Fre
 * fixed pointer/los-line part of selection
 * fixed token label has sometimes odd cut-off
 * fixed missing material on default upload preview token
+* fixed missing shadow on default upload preview token
 * refactored imports and module responsibility